Issue #5 Split up configuration of actuator endpoint and injection
